+++ src/sx-design/.storybook/SxDesignTheme.js
@@ -1,0 +1,9 @@
+// @flow
+
+import { create } from '@storybook/theming';
+
+export default (create({
+  base: 'light',
+  brandTitle: 'SX Design Storybook',
+  brandUrl: 'https://github.com/adeira/sx-design',
+}): $FlowFixMe);